pip-requirements.txt
MySQL-python==1.2.3
Django==1.3.1
django-extensions==0.8
django-reversion==1.6
#http://github.com/django-mptt/django-mptt/tarball/master#egg=django-mptt0.5
xlrd==0.7.1
xlwt==0.7.4
Babel==0.9.6
BabelDjango==0.2.2
python-dateutil==2.0
South==0.7.3
django-mptt==0.5.2
nosmsd
python-memcached
wsgiref==0.1.2
